added the closed chute block, tile, models, and textures.  Also, started work on adding multiblocks

+++ src/main/java/com/kilroy790/notenoughmachines/utilities/NEMItemHelper.java
@@ -20,14 +20,14 @@ public class NEMItemHelper {
 	public static void dropItemHandlerInventory(final IItemHandler handler, World world, BlockPos pos) {
 		if(handler == null) return;
 		for(int i = 0; i < handler.getSlots(); ++i) {
-			InventoryHelper.spawnItemStack(world, pos.getX(), pos.getY(), pos.getZ(), handler.getStackInSlot(i));
+			InventoryHelper.spawnItemStack(world, (double)pos.getX(), (double)pos.getY(), (double)pos.getZ(), handler.getStackInSlot(i));
 		}
 	}
 	
 	
 	public static void dropItemStack(World world, BlockPos pos, ItemStack stack) {
 		if(stack == null) return;
-		else InventoryHelper.spawnItemStack(world, pos.getX(), pos.getY(), pos.getZ(), stack);
+		else InventoryHelper.spawnItemStack(world, (double)pos.getX(), (double)pos.getY(), (double)pos.getZ(), stack);
 	}
